refactor(py/executor): rename find_actions to find_action_dispatch
Also resolve a yapf/isort oscillation on the linear queries and
builtins re-export imports (the failing pre-commit CI check) by
pinning the layout with an inline isort: skip (repo convention).
